requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.7.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
pypdf>=3.17.0

# Machine Learning & Training
bitsandbytes>=0.41.0
scipy>=1.10.0
scikit-learn>=1.3.0
evaluate>=0.4.0

# Web Scraping & Search
duckduckgo-search>=4.0.0
googlesearch-python>=1.2.0

# Configuration & Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Logging & Monitoring
wandb>=0.16.0
tensorboard>=2.14.0

# Text Processing
nltk>=3.8.0
spacy>=3.7.0
regex>=2023.10.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: For specific GPU support
# torch-audio>=2.0.0  # Uncomment if needed
# torchvision>=0.15.0  # Uncomment if needed 